Updated events

+++ contracts/INFTPotionWhitelist.sol
@@ -3,8 +3,8 @@ pragma solidity 0.8.9;
 
 interface INFTPotionWhitelist {
     struct WhitelistData {
-        uint256 firstId;
-        uint256 lastId;
+        uint128 firstId;
+        uint128 lastId;
     }
 
     function getWhitelistRanges(address buyer) external returns (WhitelistData[] memory);
